requirements.txt
discord.py
PyYAML
pydantic-settings